app/requirements.txt
Flask
pandas
numpy
requests
python-dotenv
htmx